moved some files for it to work
