+++ pages/api/booking.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import connectMongo from '../../lib/connectMongodb';
-import Booking from '../../models/Booking';
+import connectMongo from '../../app/lib/connectMongodb';
+import Booking from '../../app/models/Booking';
 
 export default async function handler(
   req: NextApiRequest,
@@ -23,14 +23,16 @@ export default async function handler(
       email,
       fullName,
       movieTitle,
-      movieTime: new Date(movieTime),
+      movieTime: new Date(movieTime).toLocaleString(),
       screeningId,
     });
 
     await booking.save();
     res.status(201).json({ message: 'Booking submitted succesfully', booking });
-  } catch {
-    console.error('Error submitting booking:');
-    res.status(500).json({ error: 'Failed to submit booking' });
+  } catch (error) {
+    console.error('Error submitting booking:', error.message);
+    res
+      .status(500)
+      .json({ error: `Failed to submit booking: ${error.message}` });
   }
 }
