+++ pages/boka-film/page.tsx
@@ -3,16 +3,18 @@ import BookingForm from '@/app/components/BookingForm';
 
 const BookMoviePage: React.FC = () => {
   const router = useRouter();
-  const { id, movieTitle, movieTime } = router.query;
+  const { screeningId, movieTitle, movieTime } = router.query;
+
+  if (!screeningId || !movieTitle || !movieTime) {
+    return <div>Loading...</div>;
+  }
 
   return (
     <div>
-      <h1>Book Tickets for {movieTitle}</h1>
-      <h2>Time: {movieTime}</h2>
       <BookingForm
         movieTitle={movieTitle as string}
         movieTime={new Date(movieTime as string)}
-        screeningId={id as string}
+        screeningId={screeningId as string}
       />
     </div>
   );
